requirements.txt
django
django-allauth
mysqlclient